% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_internal_r_code.R
\name{ChoosePlots}
\alias{ChoosePlots}
\title{Select plots of the dataset.}
\usage{
ChoosePlots(data, num.of.plots)
}
\arguments{
\item{data}{Complete dataset}

\item{num.of.plots}{A number}
}
\value{
chosenplots A vector with the name/number of the selected plots
}
\description{
\code{ChoosePlots} creats a vector in which every plot ID occurs once and
than selects randomly a certain number of plot IDs and saves them ordered
into a new vector.
}

