% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ecologist_r_code.R
\name{CreateEcologist}
\alias{CreateEcologist}
\title{Creates a vector with the behaviour of an ecologist.}
\usage{
CreateEcologist(sampling.area, detection.probability, identification.error,
  probability.missed.visits.forest, probability.missed.visits.farmland,
  probability.missed.visits.grassland, costs)
}
\arguments{
\item{sampling.area}{The sampled area of a cell in percent.}

\item{detection.probability}{A number in percent.}

\item{identification.error}{A number in percent.}

\item{probability.missed.visits.forest}{A number in percent.}

\item{probability.missed.visits.farmland}{A number in percent.}

\item{probability.missed.visits.grassland}{A number in percent.}

\item{costs}{A number in monetary units.}
}
\value{
A vector with the five characteristics of an ecologist.
}
\description{
\code{CreateEcologist} creates a vector with characteristics of the behaviour
of an ecologist in the field. The characteristics are: The sampled area of a
plot, the detection probability, the identification error, the probability
of missed vistis and the costs. The first four characteristics are expected
to be in percent.
}

