% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_r_code.R
\name{Evaluation}
\alias{Evaluation}
\title{Run the Model with repetitions}
\usage{
Evaluation(data, num.of.plots.forest, num.of.plots.farmland,
  num.of.plots.grassland, paid, volunteer, num.paid.forest, num.paid.farmland,
  num.paid.grassland, frequency.year, frequency.month, frequency.day, outputall,
  repetition)
}
\arguments{
\item{data}{The input dataset.}

\item{num.of.plots.forest}{Number of plots with forest habitat}

\item{num.of.plots.farmland}{Number of plots with farmland habitat}

\item{num.of.plots.grassland}{Number of plots with grassland habitat}

\item{paid}{Vector created with \code{CreateEcologist} for paid observers}

\item{volunteer}{Vector created with \code{CreateEcologist} for volunteers}

\item{num.paid.forest}{Number of plots with forest habitat that are sampled by paid observers}

\item{num.paid.farmland}{Number of plots with farmland habitat that are sampled by paid observers}

\item{num.paid.grassland}{Number of plots with grassland habitat that are sampled by paid observers}

\item{frequency.year}{Yearly sampling frequency}

\item{frequency.month}{Montly sampling frequency}

\item{frequency.day}{Daily sampling frequency}

\item{outputall}{If `True` complete dataset is returned. If False only the sampled plots are returned}

\item{repetition}{Number of repetitions of the model}
}
\value{
Result
}
\description{
\code{Evaluation} runs the 'Sampling' function with repetitions.
}

