% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_internal_r_code.R
\name{PaidPlots}
\alias{PaidPlots}
\title{Selects the plots which are monitored by paid observers.}
\usage{
PaidPlots(chosenplots, num.of.plots, num.paid)
}
\arguments{
\item{chosenplots}{Output of the function \code{ChoosePlots}}

\item{num.of.plots}{Number of sampled plots}

\item{num.paid}{Number of paid observers}
}
\value{
paid.plots A vector with the plot IDs of the cells which are sampled
by paid observers.
}
\description{
\code{PaidPlots} selects randomly the plots of the \code{chosenplots} which
are monitored by paid observers.
}

