% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ecologist_r_code.R
\name{PrepareDataset}
\alias{PrepareDataset}
\title{Rewrite the dataset into a usable format.}
\usage{
PrepareDataset(data, plot, num.of.individuals, species, x.coord, y.coord,
  visit.year, visit.month, visit.day, habitat)
}
\arguments{
\item{data}{The input dataset.}

\item{plot}{Position of the row which contains the plot or cell IDs}

\item{num.of.individuals}{Position of the row which contains the number of individuals}

\item{species}{Position of the row which contains the species names}

\item{x.coord}{Position of the row which contains the x coordinates}

\item{y.coord}{Position of the row which contains the y coordinates}

\item{visit.year}{Position of the row which contains the year date}

\item{visit.month}{Position of the row which contains the month}

\item{visit.day}{Position of the row which contains the day}

\item{habitat}{The habitat of the plot}
}
\value{
A data frame with a least 10 rows Which are in an certain order.
}
\description{
\code{PrepareDataset} rewrites the dataset to make it useable for the
Sampling function of this package. The function comibines the different
columns and creates a new dataset. If there are now data from a column the
Column will be filled with zeros.
}

