% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_internal_r_code.R
\name{PrepareSamplingEcologist}
\alias{PrepareSamplingEcologist}
\title{Prepares a subdataset for a certain combination of habitat and ecologist.}
\usage{
PrepareSamplingEcologist(data, ecologist, habitat, sampling.area,
  detection.probability, identification.error, probability.missed.visits)
}
\arguments{
\item{data}{Dataset}

\item{ecologist}{Type of ecologist which sampled the plot}

\item{habitat}{Habtiat type for which the sampling dataset should be prepared.}

\item{sampling.area}{Percentage of sampled cell area}

\item{detection.probability}{Detection probability of the species}

\item{identification.error}{Identification error of the species}

\item{probability.missed.visits}{Probability of a missed visit for the cell}
}
\value{
data Subdataset
}
\description{
\code{PrepareSamplingEcologist} selects a certain combination of plots and
ecologist which is then used in the \code{SamplingEcologist} function.
}

