% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_package_description.R
\docType{package}
\name{VirSysMon}
\alias{VirSysMon}
\alias{VirSysMon-package}
\title{VirSysMon: A package for simulating the sampling of data.}
\description{
The VirSysMon package simulates the sampling of data. It
should help to improve the results of monitoring programms or should help in
the development of them.
}
\section{VirSysMon functions}{

\itemize{
  \item PrepareDataset
  \item CreateEcologist
  \item Evaluation
  \item Sampling
}
}

\section{Intern functions}{


\itemize{
  \item ChoosePlots
  \item ExpertPlots
  \item VolunteerPlots
  \item PrepareSamplingEcologist
  \item SamplingEcologist
}
}

