% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_internal_r_code.R
\name{VolunteerPlots}
\alias{VolunteerPlots}
\title{Selects the plots which are monitored by volunteers.}
\usage{
VolunteerPlots(chosenplots, paid.plots)
}
\arguments{
\item{chosenplots}{Output of the function \code{ChoosePlots}}

\item{paid.plots}{Plots visited by paid observers}
}
\value{
volunteer.plots A vector with the name/number of the plots.
}
\description{
\code{VolunteerPlots} saves all plots of \code{chosenplots} which are not
sampled by paid observers.
}

